fix apply
