+++ src/main/crdPages/subspace/about/CrdSubspaceAboutPage.tsx
@@ -1,32 +1,22 @@
-import { Lock, Mail } from 'lucide-react';
-import { useTranslation } from 'react-i18next';
-import { useCommunityGuidelinesQuery, useSpaceAboutDetailsQuery } from '@/core/apollo/generated/apollo-hooks';
 import useNavigate from '@/core/routing/useNavigate';
-import { LoadingSpinner } from '@/crd/components/common/LoadingSpinner';
-import { CommunityGuidelinesBlock } from '@/crd/components/space/CommunityGuidelinesBlock';
-import { SpaceAboutApplyButton } from '@/crd/components/space/SpaceAboutApplyButton';
-import { SpaceAboutDialog } from '@/crd/components/space/SpaceAboutDialog';
-import type { SpaceAboutData } from '@/crd/components/space/SpaceAboutView';
 import { useSpace } from '@/domain/space/context/useSpace';
 import { useSubSpace } from '@/domain/space/hooks/useSubSpace';
-import { StorageConfigContextProvider } from '@/domain/storage/StorageBucket/StorageConfigContext';
-import { buildSettingsUrl } from '@/main/routing/urlBuilders';
-import { useSpaceApplyFlow } from '../../space/useSpaceApplyFlow';
+import { CrdSubspaceAbout } from './CrdSubspaceAbout';
 
 /**
- * CRD subspace (L1/L2) About page.
+ * CRD subspace (L1/L2) About route (`/about`).
  *
- * Counterpart to `CrdSpaceAboutPage` for subspaces. Reads from `SubspaceContext`
- * (so it shows the subspace's own about content) and uses `SpaceContext` only
- * to know the parent L0 URL for the close-fallback.
+ * Thin wrapper around the shared `CrdSubspaceAbout` — the same component the
+ * sidebar "About" trigger mounts — so both entry points render an identical
+ * About with the identical apply flow.
  *
  * Close behaviour mirrors the legacy MUI `SubspaceAboutPage`: when the viewer
- * has no read access, closing the dialog hard-navigates to the parent space URL
- * instead of going back in history — otherwise the protected-routes guard would
+ * has no read access, closing hard-navigates to the parent space URL instead
+ * of going back in history — otherwise the protected-routes guard would
  * re-fire on the previous subspace URL and trap the user in a redirect loop.
  */
 export default function CrdSubspaceAboutPage() {
-  const { subspace, permissions, loading: subspaceLoading } = useSubSpace();
+  const { subspace, permissions } = useSubSpace();
   const {
     space: {
       about: {
@@ -34,164 +24,11 @@ export default function CrdSubspaceAboutPage() {
       },
     },
   } = useSpace();
-  const { t } = useTranslation();
   const navigate = useNavigate();
 
-  const { data, loading } = useSpaceAboutDetailsQuery({
-    variables: { spaceId: subspace.id },
-    skip: !subspace.id,
-  });
-
-  const profileUrl = data?.lookup.space?.about.profile.url ?? subspace.about.profile.url;
-
-  const {
-    loading: applyLoading,
-    isMember,
-    buttonProps,
-    dialogs,
-  } = useSpaceApplyFlow({
-    spaceId: subspace.id,
-    spaceProfileUrl: profileUrl,
-    communityName: subspace.about.profile.displayName,
-  });
-
-  const guidelinesId = data?.lookup.space?.about.guidelines.id;
-  const { data: guidelinesData, loading: guidelinesLoading } = useCommunityGuidelinesQuery({
-    variables: { communityGuidelinesId: guidelinesId ?? '' },
-    skip: !guidelinesId,
-  });
-
-  if (subspaceLoading || loading) {
-    return <LoadingSpinner />;
-  }
-
-  const about = data?.lookup.space?.about;
-  if (!about) {
-    return null;
-  }
-
-  const profile = about.profile;
-
-  const leadUsers = (about.membership?.leadUsers ?? [])
-    .filter((u): u is typeof u & { profile: NonNullable<typeof u.profile> } => !!u.profile)
-    .map(u => ({
-      name: u.profile.displayName,
-      avatarUrl: u.profile.avatar?.uri,
-      type: 'person' as const,
-      location: [u.profile.location?.city, u.profile.location?.country].filter(Boolean).join(', ') || undefined,
-      href: u.profile.url,
-    }));
-
-  const leadOrgs = (about.membership?.leadOrganizations ?? [])
-    .filter((o): o is typeof o & { profile: NonNullable<typeof o.profile> } => !!o.profile)
-    .map(o => ({
-      name: o.profile.displayName,
-      avatarUrl: o.profile.avatar?.uri,
-      type: 'organization' as const,
-      location: [o.profile.location?.city, o.profile.location?.country].filter(Boolean).join(', ') || undefined,
-      href: o.profile.url,
-    }));
-
-  const provider = about.provider?.profile
-    ? {
-        name: about.provider.profile.displayName,
-        avatarUrl: about.provider.profile.avatar?.uri,
-        type: 'organization' as const,
-        location:
-          [about.provider.profile.location?.city, about.provider.profile.location?.country]
-            .filter(Boolean)
-            .join(', ') || undefined,
-        href: about.provider.profile.url,
-      }
-    : undefined;
-
-  const aboutData: SpaceAboutData = {
-    name: profile.displayName,
-    tagline: profile.tagline ?? undefined,
-    description: profile.description ?? undefined,
-    location: [profile.location?.city, profile.location?.country].filter(Boolean).join(', ') || undefined,
-    metrics: (about.metrics ?? []).map(m => ({ name: m.name, value: m.value })),
-    who: about.who ?? undefined,
-    why: about.why ?? undefined,
-    provider,
-    leadUsers,
-    leadOrganizations: leadOrgs,
-    references: (profile.references ?? []).map(r => ({
-      name: r.name,
-      uri: r.uri,
-      description: r.description ?? undefined,
-    })),
-  };
-
-  const whyTitle = t(`about.context.${subspace.level}.why` as const, { ns: 'crd-space' });
-  const whoTitle = t(`about.context.${subspace.level}.who` as const, { ns: 'crd-space' });
-
-  const lockTooltipSlot = !permissions.canRead ? (
-    <span
-      role="img"
-      aria-label={t('about.lockTooltip', { ns: 'crd-space' })}
-      className="inline-flex items-center justify-center text-primary"
-    >
-      <Lock className="w-4 h-4" aria-hidden="true" />
-    </span>
-  ) : undefined;
-
-  const showApplyButton = !isMember && !applyLoading;
-  const joinSlot = showApplyButton ? <SpaceAboutApplyButton {...buttonProps} /> : undefined;
-
-  const contactHostSlot = provider ? (
-    <a
-      href={provider.href}
-      className="inline-flex items-center gap-2 text-body-emphasis text-primary hover:underline underline-offset-4"
-    >
-      <Mail className="w-4 h-4 shrink-0" aria-hidden="true" />
-      {t('about.contactHost', { ns: 'crd-space' })}
-    </a>
-  ) : undefined;
-
-  const guidelines = guidelinesData?.lookup.communityGuidelines?.profile;
-  const guidelinesSlot = guidelinesId ? (
-    <CommunityGuidelinesBlock
-      displayName={guidelines?.displayName}
-      description={guidelines?.description ?? undefined}
-      references={guidelines?.references?.map(r => ({
-        name: r.name,
-        uri: r.uri,
-        description: r.description ?? undefined,
-      }))}
-      loading={guidelinesLoading}
-      canEdit={permissions.canUpdate}
-      onEditClick={() => navigate(`${buildSettingsUrl(profileUrl)}/community`)}
-    />
-  ) : undefined;
-
   const handleClose = () => {
-    navigate(permissions.canRead ? profileUrl : parentSpaceUrl);
+    navigate(permissions.canRead ? subspace.about.profile.url : parentSpaceUrl);
   };
 
-  return (
-    <StorageConfigContextProvider locationType="space" spaceId={subspace.id}>
-      <SpaceAboutDialog
-        open={true}
-        onOpenChange={open => {
-          if (!open) {
-            handleClose();
-          }
-        }}
-        data={aboutData}
-        hasEditPrivilege={permissions.canUpdate}
-        whyTitle={whyTitle}
-        whoTitle={whoTitle}
-        lockTooltipSlot={lockTooltipSlot}
-        joinSlot={joinSlot}
-        guidelinesSlot={guidelinesSlot}
-        contactHostSlot={contactHostSlot}
-        onEditDescription={() => navigate(`${buildSettingsUrl(profileUrl)}/about#description`)}
-        onEditWhy={() => navigate(`${buildSettingsUrl(profileUrl)}/about#why`)}
-        onEditWho={() => navigate(`${buildSettingsUrl(profileUrl)}/about#who`)}
-        onEditReferences={() => navigate(`${buildSettingsUrl(profileUrl)}/about`)}
-      />
-      {dialogs}
-    </StorageConfigContextProvider>
-  );
+  return <CrdSubspaceAbout open={true} onClose={handleClose} />;
 }
